Revert "FLEET-9992: Update job carrier location type for use with fleet services"
This reverts commit d6bb3e1f0141f2c873b17fd08d889c36720b9cf9.

+++ MiX.Integrate.Shared/Entities/Messages/SendJobMessageCarrier.cs
@@ -19,7 +19,7 @@ namespace MiX.Integrate.Shared.Entities.Messages
 		public bool AddAddressSummary { get; set; }
 		public bool UseFirstAddressForSummary { get; set; }
 		public JobMessageActionNotifications NotificationSettings { get; set; }
-		public long[] AddressList { get; set; }
+		public int[] AddressList { get; set; }
 		public CommsTransports Transport { get; set; }
 		public bool Urgent { get; set; }
 	}
